Started shimFS: Cacheing works.
The purpose of ShimFS is to act as a cache'ing and async'ing layer for
any MinFS (might be useful for remote filesystems). This initial version
cache's readdirectory and stat results for 5 seconds, and cache's file
contents till the computer runs out of memory and crashes :).   For now
it's Synced, but it'll probably be turned Async where possible
(writefile, etc.) in the future.