Use standard errors; Streamline GET/SETATTR, CREATE
When possible, it's best to use the standard errors found in the os
package in golang, and the NFS3 errors in C code.

Continuing with the streamlining, I've looked over GETATTR, SETATTR, and
CREATE.

Also, I decided to simplify MinFS.Remove() by removing the recursive
option.  I don't really need it for my use case, and it would've been a
pain to get working right in all circumstances and backends. Also, most
OS's and programs are willing to delve into and empty directories before
being able to delete them.