Limit shimfs to FileInfo & Readdir data. Bugfixes.
I'm limiting shimfs to FileInfo and Readdir data for now (i.e. I'm
ripping out file data cacheing for now). This is just to give me a
chance to make sure everything else is working correctly, and to later
re-implement data cacheing in a saner and easier to keep track of way.
Default timeout is now set to 5 seconds.

Also, a few minor bug fixes with NFS Stat and Commit handling, as well
as bad CLI input.